Support union object stores for IDBPObjectStore (with caveats)
Something weird is going on with the function return types. The `Value` type parameter seems to be different depending on whether it's in the parameter type context or the return type context.

Also, removed `IDBKeyRange` from the `add` and `put` methods because that seems to be a mistake. It's not included in lib.dom.d.ts.